Use default wake interval for WorkerPool
Currently Interval defaults to unset/zero which will result in
worker pools that spin and peg all available cycles querying for new
jobs. This patch changes the default interval to use the same logic
for NewWorkerPool as NewWorker and use a default of
QUE_WAKE_INTERVAL or five seconds if unset.

Signed-off-by: Jonathan Rudenberg <jonathan@titanous.com>