Support connections, pools, and transactions for statement preparation
pgx supports preparing statements at any of these three levels, ensuring
that pool prepared statements will apply to all current and future
connections, and running statement preparation on a transaction's
underlying connection.

Introduce a Preparer interface that these three types all
satisfy, and use that as the argument for the new PrepareStatementsWithPreparer
func. This allows users of this package to use PrepareStatements as the
AfterConnect option when creating a pool, or to pass an existing pool  or
transaction into PrepareStatementsWithPreparer.